requirements.txt
django
djangorestframework
psycopg2-binary
celery
requests